Rename control_plane::provider to control_plane::client
